refactor(poke-auth-service): rename it to user-auth-service
